Add headless rendering mode, and run WPT tests with Webrender enabled.
When -z is specified, create a glutin headless rendering context. On
Linux, this uses OSMesa to do software rendering without any dependency
on X or GPU hardware.

This requires a newer Mesa than is present on most Linux distributions,
so Mesa is built from source as a cargo dependency.

Switch the default WPT test mode to Webrender.